requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
PyYAML>=6.0.1
PyPDF2>=3.0.1
numpy>=1.26.0
tqdm>=4.66.0
rank-bm25>=0.2.2
jinja2>=3.1.3
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
transformers>=4.51.0
torch>=2.2.0
requests>=2.31.0
vllm>=0.5.0